Test program startrek.bas
I ported the startrek program over so that it will run under the current richpl version PyBasic. I threw together a quick python script to separate the compound lines which caused a few line number bugs, but I think I found most of those. Since PyBasic doesn't yet support the trailing delimiter to suppress the end of line carriage return some of the displays are off but I did rebuild the printouts of most of the primary control displays.